+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ not yet released
 ----------------------------------
 
 * #161: Added Swedish translation.
+* #xxx: Added support for FreeVoipDeal SMS gateway <http://www.freevoipdeal.com>
 
 
 v3.1.1 (released November 6, 2014)